+++ ChangeLog.md
@@ -76,6 +76,9 @@
 
 - Removes `PureC` in favour of `Data.Functor.Identity.Identity`. Note that `run` is still provided as a synonym for `runIdentity`. ([#305](https://github.com/fused-effects/fused-effects/pull/305))
 
+- Redefines `ReaderC` as a type & pattern synonym of `ReaderT`. ([#305](https://github.com/fused-effects/fused-effects/pull/305))
+
+
 # v0.5.0.1
 
 - Adds support for ghc 8.8.1.